+++ R/fun-format.R
@@ -161,6 +161,7 @@
 #'
 #' @export
 #' @examples
+#' old_log <- Sys.getenv("SDC_LOG_LEVEL")
 #' Sys.setenv(SDC_LOG_LEVEL = "OFF")
 #' dt <- data.table::data.table(
 #'   country = sample(c("AT", "DE", "NL"), 100, replace = TRUE),
@@ -176,6 +177,7 @@
 #'   rn_rebalance(dim_list = dims, num_var = "turnover") |>
 #'   rn_perturb(dim_list = dims, variables = "turnover")
 #' rn_format(res, format = "long")
+#' Sys.setenv(SDC_LOG_LEVEL = old_log)
 rn_format <- function(x, format = "wide", variables = NULL) {
   .assert_rn_class(
     x,