+++ R/rebalancedNoise-package.R
@@ -2,14 +2,11 @@
 #' @import R6
 #' @import data.table
 #' @import sdcTable
-#' @import progress
 #' @useDynLib rebalancedNoise, .registration = TRUE
 #' @importFrom Rcpp sourceCpp
 #' @importFrom sdcHierarchies hier_create
 #' @importFrom parallel detectCores
 #' @importFrom stats quantile
-#' @importFrom Rcpp sourceCpp
-#' @useDynLib rebalancedNoise, .registration = TRUE
 "_PACKAGE"
 
 utils::globalVariables(c(