+++ R/zzz.R
@@ -1,4 +1,4 @@
 .onAttach <- function(libname, pkgname) {
   v <- utils::packageVersion(pkgname)
-  packageStartupMessage(glue::glue("Welcome to rebalancedNoise {v}."))
+  packageStartupMessage("Welcome to rebalancedNoise ", v, ".")
 }